Format.

+++ src/pke/include/schemebase/base-scheme.h
@@ -1560,7 +1560,7 @@ public:
 
     void EvalBootstrapSetup(const CryptoContextImpl<Element>& cc, const std::vector<uint32_t>& levelBudget = {5, 4},
                             const std::vector<uint32_t>& dim1 = {0, 0}, uint32_t slots = 0,
-                            uint32_t correctionFactor = 0.0) {
+                            double correctionFactor = 0.0) {
         if (m_FHE) {
             m_FHE->EvalBootstrapSetup(cc, levelBudget, dim1, slots, correctionFactor);
             return;
